Manual File Watching.
Removed dependency on nodemon and replaced it with node-watch.
Manually fork the './entry-server-instance' which allows for using the
built-in ipc between parent and child.

Child will now send messages when it shutsdown and is ready to serve
requests, which allows the child to wait until the server is ready
before it attempts to proxy requests to it.